fix(coderd/x/chatd): use documented Bedrock body cap
